django/requirements.txt
DateUtils==0.5.2
Django==1.4.3
MySQL-python==1.2.3
PIL==1.1.7
South==0.7.6
distribute==0.6.24
django-debug-toolbar==0.9.4
django-filebrowser==3.5.1
django-grappelli==2.4.3
django-tinymce==1.5.1b4
python-dateutil==2.1
django-crispy-forms==1.2.3
pytz==2012c
six==1.2.0
vobject==0.8.1c
wsgiref==0.1.2
